Remove warnings from generated adaptors.
This patch modifies the dbusxx-xml2cpp's generate_adaptor function  so
that it prevents generation of the MessageIter ri in cases where an
interface method has no input arguments.

By making this element conditional based on the presence of input
arguments we avoid "unused-but-set-variable" warning in code that uses
the DBus-C++ adaptor glue headers.

Change-Id: I0f04439bf7b2b3cecb9d639e61c2b6dedc148e51
Type: Code Improvement